+++ Backend/src/main/java/com/github/backend/web/admin/AdminController.java
@@ -3,6 +3,9 @@ package com.github.backend.web.admin;
 import com.github.backend.dto.admin.LoginAdmin;
 import com.github.backend.dto.admin.RegisterAdminOutputDto;
 import com.github.backend.dto.common.AdminInfo;
+import com.github.backend.dto.course.SearchCourse;
+import com.github.backend.dto.member.MemberSearchOutputDto;
+import com.github.backend.dto.member.SearchMember;
 import com.github.backend.security.jwt.Tokens;
 import com.github.backend.dto.course.CourseInfoOutputDto;
 import com.github.backend.dto.admin.CreateAdmin;
@@ -46,35 +49,7 @@ public class AdminController {
     }
 
 
-    @PostMapping("/admin/login")
-    public ResponseEntity<Result<?>> loginAdmin(
-            @RequestBody @Valid LoginAdmin.Request request
-    ){
-        Tokens tokens = adminService.loginAdmin(
-                request.getAdminId(), request.getPassword()
-        );
 
-        // convert Entity to DTO
-        return ResponseEntity.ok().body(
-                Result.builder()
-                        .status(200)
-                        .success(true)
-                        .data(tokens)
-                        .build()
-        );
-    }
-
-    @GetMapping("/admin/logout")
-    public ResponseEntity<Result<?>> logout(@AuthenticationPrincipal AdminInfo adminInfo){
-
-        return ResponseEntity.ok().body(
-                Result.builder()
-                        .status(200)
-                        .success(true)
-                        .data(adminService.logoutAdmin(adminInfo))
-                        .build()
-        );
-    }
 
     @PostMapping("/gamer")
     public ResponseEntity<Result<?>> registerGamer(
@@ -230,5 +205,6 @@ public class AdminController {
                         .build()
         );
     }
+
 
 }
